# swagger.yml
openapi: 3.0.1
info:
  title: API Gateway service API endpoints
  description: API Gateway from Message in a Bottle with Microservices
  version: "0.1"
servers:
- url: /
paths:
...
  /user/:
    get:
      tags: [ User ]
      summary: This method allows to retrieve the sign up page for a new user
      parameters:
        - name: type_
          description: type of the user
          in: path
          required: true
          schema: 
            type: string
      responses:
        200:
          description: Renders the create_user template, filled with the fiels for the user's data
    post:
      summary: Add a new customer
      tags: [ User ]
      parameters:
        - name: type_
          description: type of the user
          in: path
          required: true
          schema: 
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: the email field for the user
                password:
                  type: string
                  description: the password field for the user
                social_number:
                  type: string
                  description: the social number field for the user
                firstname: 
                  type: string
                  description: the firstname field for the user
                lastname: 
                  type: string
                  description: the lastname field for the user
                dateofbirth:
                  type: string
                  description: the birthday field for the user
              required:
                - email
                - password
      responses:
        302:
          description: Redirects to the correct profile page, depending on the type of the user
...